Added initial version of the base ShadowBot plugin.
At the moment this plugin consists of nothing more than an extension of events.EventEmitter. Using this class as a base for all plugins allows us to check programmatically if a plugin provided to ShadowBot is actually a plugin. This can later be used to provide helper methods and properties to plugins.